+++ clean_python/base/domain/__init__.py
@@ -1,3 +1,4 @@
+from .context import *  # NOQA
 from .domain_event import *  # NOQA
 from .domain_service import *  # NOQA
 from .exceptions import *  # NOQA