+++ clean_python/fastapi/__init__.py
@@ -1,6 +1,6 @@
-from .context import *  # NOQA
 from .error_responses import *  # NOQA
 from .fastapi_access_logger import *  # NOQA
 from .request_query import *  # NOQA
 from .resource import *  # NOQA
+from .security import *  # NOQA
 from .service import *  # NOQA